Refactor platform specific code
