+++ pkg/crc/machine/driver_darwin.go
@@ -2,7 +2,12 @@ package machine
 
 import (
 	"github.com/code-ready/crc/pkg/crc/constants"
+	"github.com/code-ready/crc/pkg/crc/errors"
 	crcos "github.com/code-ready/crc/pkg/os"
+
+	"github.com/code-ready/crc/pkg/crc/machine/config"
+	"github.com/code-ready/crc/pkg/crc/machine/hyperkit"
+	"github.com/code-ready/crc/pkg/crc/machine/virtualbox"
 )
 
 func init() {
@@ -25,4 +30,22 @@ func init() {
 	}
 
 	DefaultDriver = HyperkitDriver
+}
+
+func getDriverOptions(machineConfig config.MachineConfig) interface{} {
+	var driver interface{}
+
+	// Supported drivers
+	switch machineConfig.VMDriver {
+
+	case "virtualbox":
+		driver = virtualbox.CreateHost(machineConfig)
+	case "hyperkit":
+		driver = hyperkit.CreateHost(machineConfig)
+
+	default:
+		errors.ExitWithMessage(1, "Unsupported driver: %s", machineConfig.VMDriver)
+	}
+
+	return driver
 }
