+++ pkg/crc/machine/driver_linux.go
@@ -2,7 +2,11 @@ package machine
 
 import (
 	"github.com/code-ready/crc/pkg/crc/constants"
+	"github.com/code-ready/crc/pkg/crc/errors"
 	crcos "github.com/code-ready/crc/pkg/os"
+
+	"github.com/code-ready/crc/pkg/crc/machine/config"
+	"github.com/code-ready/crc/pkg/crc/machine/libvirt"
 )
 
 func init() {
@@ -18,4 +22,20 @@ func init() {
 	}
 
 	DefaultDriver = LibvirtDriver
+}
+
+func getDriverOptions(machineConfig config.MachineConfig) interface{} {
+	var driver interface{}
+
+	// Supported drivers
+	switch machineConfig.VMDriver {
+
+	case "libvirt":
+		driver = libvirt.CreateHost(machineConfig)
+
+	default:
+		errors.ExitWithMessage(1, "Unsupported driver: %s", machineConfig.VMDriver)
+	}
+
+	return driver
 }
